Extract link preview core
